Fix UB in NetworkMessage and race conditions in NetworkConnection
- Extracted NetworkMessage to separate files (network_message.h/cpp).
- Added bounds checking to NetworkMessage::read<T> and read<std::string> to prevent buffer overflows (UB).
- Used std::memcpy in NetworkMessage::read<T> to prevent strict aliasing violations and unaligned access (UB).
- Changed NetworkConnection::stopped to std::atomic<bool> to fix data race (UB).
- Added std::mutex to NetworkConnection to synchronize start/stop operations and prevent race conditions.
- Updated CMakeLists.txt to include new network_message.cpp.
