+++ source/net/net_connection.cpp
@@ -19,57 +19,6 @@
 #include "net/net_connection.h"
 #include <spdlog/spdlog.h>
 
-NetworkMessage::NetworkMessage() {
-	clear();
-}
-
-void NetworkMessage::clear() {
-	buffer.resize(4);
-	position = 4;
-	size = 0;
-}
-
-void NetworkMessage::expand(const size_t length) {
-	if (position + length >= buffer.size()) {
-		buffer.resize(position + length + 1);
-	}
-	size += length;
-}
-
-template <>
-std::string NetworkMessage::read<std::string>() {
-	const uint16_t length = read<uint16_t>();
-	char* strBuffer = reinterpret_cast<char*>(&buffer[position]);
-	position += length;
-	return std::string(strBuffer, length);
-}
-
-template <>
-Position NetworkMessage::read<Position>() {
-	Position position;
-	position.x = read<uint16_t>();
-	position.y = read<uint16_t>();
-	position.z = read<uint8_t>();
-	return position;
-}
-
-template <>
-void NetworkMessage::write<std::string>(const std::string& value) {
-	const size_t length = value.length();
-	write<uint16_t>(length);
-
-	expand(length);
-	memcpy(&buffer[position], &value[0], length);
-	position += length;
-}
-
-template <>
-void NetworkMessage::write<Position>(const Position& value) {
-	write<uint16_t>(value.x);
-	write<uint16_t>(value.y);
-	write<uint8_t>(value.z);
-}
-
 // NetworkConnection
 NetworkConnection::NetworkConnection() :
 	service(nullptr), thread(), stopped(false) {
@@ -86,6 +35,8 @@ NetworkConnection& NetworkConnection::getInstance() {
 }
 
 bool NetworkConnection::start() {
+	std::lock_guard<std::mutex> lock(connection_mutex);
+
 	if (thread.joinable()) {
 		if (stopped) {
 			return false;
@@ -113,6 +64,8 @@ bool NetworkConnection::start() {
 }
 
 void NetworkConnection::stop() {
+	std::unique_lock<std::mutex> lock(connection_mutex);
+
 	if (!service) {
 		return;
 	}
@@ -124,11 +77,13 @@ void NetworkConnection::stop() {
 	stopped = true;
 
 	if (thread.joinable()) {
+		lock.unlock();
 		spdlog::info("NetworkConnection::stop - joining thread");
 		spdlog::default_logger()->flush();
 		thread.join();
 		spdlog::info("NetworkConnection::stop - thread joined");
 		spdlog::default_logger()->flush();
+		lock.lock();
 	}
 
 	service.reset();
